services/server-python/requirements.txt
# cycler==0.10.0
grpcio==1.35.0
grpcio-tools==1.35.0
# kiwisolver==1.3.1
# matplotlib==3.3.4
# numpy==1.20.1
# Pillow==8.1.1
protobuf==3.15.4
# pyparsing==2.4.7
# python-dateutil==2.8.1
# six==1.15.0
